tidy up, move shims/ folder into src/
