+++ .npmignore
@@ -4,4 +4,3 @@
 # but include these folders
 !/build/**/*
 !/src/**/*
-!/shims/**/*
