+++ src/Ros3D.js
@@ -1,4 +1,4 @@
-import THREE from '../shims/three/core.js';
+import THREE from './shims/three/core.js';
 
 /**
  * @author Russell Toris - rctoris@wpi.edu
