+++ src/markers/MarkerArrayClient.js
@@ -1,6 +1,6 @@
-import THREE from '../../shims/three/core.js';
-import { EventEmitter2 } from '../../shims/eventemitter2/EventEmitter2.js';
-import ROSLIB from '../../shims/roslib/ROSLIB.js';
+import THREE from '../shims/three/core.js';
+import { EventEmitter2 } from '../shims/eventemitter2/EventEmitter2.js';
+import ROSLIB from '../shims/roslib/ROSLIB.js';
 
 import { Marker } from './Marker.js';
 import { SceneNode } from '../visualization/SceneNode.js';
