+++ src/models/Arrow2.js
@@ -1,4 +1,4 @@
-import THREE from '../../shims/three/core.js';
+import THREE from '../shims/three/core.js';
 
 /**
  * @author Jihoon Lee - lee@magazino.eu
