+++ src/navigation/OccupancyGridClient.js
@@ -1,6 +1,6 @@
-import { EventEmitter2 } from '../../shims/eventemitter2/EventEmitter2.js';
-import THREE from '../../shims/three/core.js';
-import ROSLIB from '../../shims/roslib/ROSLIB.js';
+import { EventEmitter2 } from '../shims/eventemitter2/EventEmitter2.js';
+import THREE from '../shims/three/core.js';
+import ROSLIB from '../shims/roslib/ROSLIB.js';
 
 import { OccupancyGrid } from './OccupancyGrid.js';
 import { SceneNode } from '../visualization/SceneNode.js';
