+++ src/shims/eventemitter2/EventEmitter2.js
