+++ src/shims/three/ColladaLoader.js
