+++ src/shims/three/STLLoader.js
