+++ src/shims/three/core.js
