+++ src/urdf/Urdf.js
@@ -1,5 +1,5 @@
-import THREE from '../../shims/three/core.js';
-import ROSLIB from '../../shims/roslib/ROSLIB.js';
+import THREE from '../shims/three/core.js';
+import ROSLIB from '../shims/roslib/ROSLIB.js';
 
 import { makeColorMaterial } from '../Ros3D.js';
 import { MeshLoader } from '../models/MeshLoader.js';
