+++ test/depthCloud/DepthCloud.test.js
@@ -1,3 +1,5 @@
+/* globals chai */
+
 var assert = chai.assert;
 
 describe('depthCloud', function() {
